refactor-3 Pass 3C-3: extract ActionsWindowUI from ParsekUI
Move actions window into dedicated UI/ActionsWindowUI.cs class.
19 fields, ActionsSortColumn enum, and 9 methods moved. ParsekUI
delegates via actionsUI.DrawIfOpen() and toggle via IsOpen property.
